+++ src/model/allIngestJobsInfoResponse.ts
@@ -1,0 +1,7 @@
+import { JobUploadResponse } from './jobUploadResponse';
+
+export interface AllIngestJobsInfoResponse {
+    done: boolean;
+    records: JobUploadResponse[];
+    nextRecordsUrl: string;
+}