+++ src/model/bulkJobInfoResponse.ts
@@ -3,5 +3,6 @@ import { BulkQueryResponse } from "./queryResponse";
 export interface BulkJobInfoResponse extends BulkQueryResponse {
     numberRecordsProcessed : number,
     retries: number,
-    totalProcessingTime: number
+    totalProcessingTime: number,
+    isPkChunkingSupported?: boolean
 }