+++ src/model/connection.ts
@@ -1,6 +1,10 @@
 export interface Connection {
     accessToken: string;
     apiVersion: string;
-    instanceUrl: string
-    isTooling?: boolean
+    instanceUrl: string;
+    isTooling?: boolean;
+    callOptions?: {
+        client?: string;
+        defaultNamespace?: string;
+    }
 }