+++ src/model/ingestJobConfig.ts
@@ -1,0 +1,5 @@
+export interface IngestJobConfig {
+    isPkChunkingEnabled?: string;
+    jobType?: string;
+    queryLocator?: number;
+}