+++ src/model/parallelQueryResultsResponse.ts
@@ -1,0 +1,9 @@
+export interface ResultPage {
+    resultUrl: string;
+}
+
+export interface ParallelQueryResultsResponse {
+    resultPages: ResultPage[];
+    nextRecordsUrl: string;
+    done: boolean;
+}